Add Grammers in Parser

+++ HardcoreDDLC/Actions/DDLCMoveAction.cs
@@ -10,11 +10,11 @@ namespace HardcoreDDLC.Actions
     class DDLCMoveAction : DDLCAction
     {
         public DependencyObject Object { get; set; }
-        public int Speed { get; set; }
+        public double Speed { get; set; }
         public Point Position { get; set; }
         public bool Active { get; set; }
 
-        public DDLCMoveAction(DependencyObject dependencyObject, Point pos, int speed = 0)
+        public DDLCMoveAction(DependencyObject dependencyObject, Point pos, double speed = 0)
         {
             Object = dependencyObject;
             Position = pos;
